Implemented profile pictures

+++ back-end/src/account/account.resolver.ts
@@ -74,7 +74,9 @@ export class AccountResolver {
   ): Promise<User> {
     let user = await User.findOne(userId);
 
-    await this.fileService.deletePublicImageFromUrl(user.image);
+    if (user.image) {
+      await this.fileService.deletePublicImageFromUrl(user.image);
+    }
     const imageUrl = await this.fileService.storePublicImageFromBase64(
       newImage,
     );
